UI: Improved Intel UI (added Grid vs. List toggle), Created new SettingsUI and adjusted SettingsForm to use it.
Intel:
Improved the IntelPresentation for text display
Improved the Intel parsing (will now grab the last word in the Intel, as long as it is not part of the solar system name)
Added option to skip the older Intel, instead of stopping after the first 5 - still has an issue of reading the very first Intel, due to the race condition.
Added Logic to detect 'player is docked' Intel.